Conflicts:
	picknik_main/launch/camera_calibration.launch

+++ picknik_main/src/execution_interface.cpp
@@ -55,7 +55,7 @@ ExecutionInterface::ExecutionInterface(bool verbose, RemoteControlPtr remote_con
                                        moveit_grasps::GraspDatas grasp_datas,
                                        planning_scene_monitor::PlanningSceneMonitorPtr planning_scene_monitor,
                                        ManipulationDataPtr config, const std::string& package_path, 
-                                       moveit::core::RobotStatePtr current_state)
+                                       moveit::core::RobotStatePtr current_state, bool fake_execution)
   : verbose_(verbose)
   , remote_control_(remote_control)
   , visuals_(visuals)
@@ -65,6 +65,8 @@ ExecutionInterface::ExecutionInterface(bool verbose, RemoteControlPtr remote_con
   , package_path_(package_path)
   , current_state_(current_state)
   , nh_("~")
+  , unit_testing_enabled_(false)
+  , fake_execution_(fake_execution)
 {
   // Check that controllers are ready
   zaber_list_controllers_client_ = nh_.serviceClient<controller_manager_msgs::ListControllers>("/jacob/zaber/controller_manager/list_controllers");
@@ -73,7 +75,8 @@ ExecutionInterface::ExecutionInterface(bool verbose, RemoteControlPtr remote_con
   ROS_INFO_STREAM_NAMED("execution_interface","ExecutionInterface Ready.");
 }
 
-bool ExecutionInterface::executeTrajectory(moveit_msgs::RobotTrajectory &trajectory_msg, bool ignore_collision)
+bool ExecutionInterface::executeTrajectory(moveit_msgs::RobotTrajectory &trajectory_msg, 
+                                           const robot_model::JointModelGroup* jmg, bool ignore_collision)
 {
   trajectory_msgs::JointTrajectory& trajectory = trajectory_msg.joint_trajectory;
 
@@ -96,24 +99,23 @@ bool ExecutionInterface::executeTrajectory(moveit_msgs::RobotTrajectory &traject
     if (trajectory.joint_names.size() > 3)
     {
       static std::size_t trajectory_count = 0;
-      saveTrajectory(trajectory_msg, "trajectory_"+ boost::lexical_cast<std::string>(trajectory_count++)+".csv");
+      saveTrajectory(trajectory_msg, jmg->getName() + "_trajectory_"+ boost::lexical_cast<std::string>(trajectory_count++)+".csv");
       //saveTrajectory(trajectory_msg, "trajectory.csv");
     }
   }
 
   // Visualize the hand/wrist path in Rviz
-  if (trajectory.points.size() > 1)
+  if (trajectory.points.size() > 1 && !jmg->isEndEffector())
   {
     visuals_->trajectory_lines_->deleteAllMarkers();
-    ros::spinOnce();
+    ros::spinOnce(); // TODO remove?
     
-    //const moveit::core::LinkModel* robot_link = planning_scene_monitor_->getRobotModel()->getLinkModel("jaco2_link_finger_2_tip");
-    visuals_->trajectory_lines_->publishTrajectoryLine(trajectory_msg, grasp_datas_[config_->right_arm_]->parent_link_,
-                                                 config_->right_arm_, rvt::LIME_GREEN);
+    visuals_->trajectory_lines_->publishTrajectoryLine(trajectory_msg, grasp_datas_[jmg]->parent_link_,
+                                                       config_->right_arm_, rvt::LIME_GREEN);
   }
   else
     ROS_WARN_STREAM_NAMED("execution_interface","Not visualizing path because trajectory only has "
-                          << trajectory.points.size() << " points");
+                          << trajectory.points.size() << " points or because is end effector");
 
   // Visualize trajectory in Rviz
   bool wait_for_trajetory = false;
@@ -136,6 +138,10 @@ bool ExecutionInterface::executeTrajectory(moveit_msgs::RobotTrajectory &traject
       std::cout << "Diff time: " << diff.toSec() << std::endl;
       std::cout << "-------------------------------------------------------" << std::endl;
       std::cout << std::endl;
+
+      remote_control_->setAutonomous(false);
+      remote_control_->setFullAutonomous(false);
+      
       //return false;
     }
     else if (diff > warn_time_step)
@@ -157,16 +163,23 @@ bool ExecutionInterface::executeTrajectory(moveit_msgs::RobotTrajectory &traject
     plan_execution_.reset(new plan_execution::PlanExecution(planning_scene_monitor_, trajectory_execution_manager_));
   }
 
+  // Check if in unit testing mode
+  if (unit_testing_enabled_)
+  {
+    robot_trajectory::RobotTrajectoryPtr 
+      robot_trajectory(new robot_trajectory::RobotTrajectory(planning_scene_monitor_->getRobotModel(), jmg));
+                                                             
+    robot_trajectory->setRobotTrajectoryMsg(*current_state_, trajectory_msg);
+    *current_state_ = robot_trajectory->getLastWayPoint();
+    return true;
+  }
+
   // Confirm trajectory before continuing
   if (!remote_control_->getFullAutonomous() &&
       // Only wait for non-finger trajectories
       trajectory.joint_names.size() > 3)
   {
-    std::cout << std::endl;
-    std::cout << std::endl;
-
-    ROS_INFO_STREAM_NAMED("execution_interface","Waiting before executing trajectory");
-    remote_control_->waitForNextFullStep();
+    remote_control_->waitForNextFullStep("execute trajectory");
     ROS_INFO_STREAM_NAMED("execution_interface","Executing trajectory....");
   }
 
@@ -247,23 +260,23 @@ bool ExecutionInterface::checkExecutionManager()
   }
 
   // Check that correct controllers are running
-  // bool has_error = true;
+  bool has_error = true;
 
-  // while (has_error && ros::ok())
-  // {
-  //   has_error = false;
-  //   if (!checkTrajectoryController(zaber_list_controllers_client_, "zaber"))
-  //   {
-  //     has_error = true;
-  //   }
+  while (has_error && ros::ok())
+  {
+    has_error = false;
+    if (!checkTrajectoryController(zaber_list_controllers_client_, "zaber"))
+    {
+      has_error = true;
+    }
 
-  //   bool has_ee = true;
-  //   if (!checkTrajectoryController(kinova_list_controllers_client_, "kinova", has_ee))
-  //   {
-  //     has_error = true;
-  //   }
-  //   ros::Duration(0.5).sleep();
-  // }
+    bool has_ee = true;
+    if (!checkTrajectoryController(kinova_list_controllers_client_, "kinova", has_ee))
+    {
+      has_error = true;
+    }
+    ros::Duration(0.5).sleep();
+  }
 
   return true;
 }
@@ -275,17 +288,20 @@ bool ExecutionInterface::checkTrajectoryController(ros::ServiceClient& service_c
   controller_manager_msgs::ListControllers service;
   if (!service_client.call(service))
   {
-    ROS_ERROR_STREAM_NAMED("execution_interface","Unable to check if controllers for " << hardware_name << " are loaded, failing. Using nh namespace " << nh_.getNamespace());
-    std::cout << "service: " << service.response << std::endl;
+    ROS_ERROR_STREAM_THROTTLE_NAMED(2, "execution_interface","Unable to check if controllers for " << hardware_name 
+                                    << " are loaded, failing. Using nh namespace " << nh_.getNamespace() 
+                                    << ". Service response: " << service.response);
     return false;
   }
+
+  std::string control_type = fake_execution_ ? "position" : "velocity";
 
   // Check if proper controller is running
   bool found_main_controller = false;
   bool found_ee_controller = false;
   for (std::size_t i = 0; i < service.response.controller.size(); ++i)
   {
-    if (service.response.controller[i].name == "velocity_trajectory_controller")
+    if (service.response.controller[i].name == control_type + "_trajectory_controller")
     {
       found_main_controller = true;
       if (service.response.controller[i].state != "running")
@@ -294,7 +310,7 @@ bool ExecutionInterface::checkTrajectoryController(ros::ServiceClient& service_c
         return false;
       }
     }
-    if (service.response.controller[i].name == "ee_velocity_trajectory_controller")
+    if (service.response.controller[i].name == "ee_" + control_type + "_trajectory_controller")
     {
       found_ee_controller = true;
       if (service.response.controller[i].state != "running")
@@ -307,12 +323,12 @@ bool ExecutionInterface::checkTrajectoryController(ros::ServiceClient& service_c
 
   if (has_ee && !found_ee_controller)
   {
-    ROS_ERROR_STREAM_THROTTLE_NAMED(2, "execution_interface","No end effector controller found for " << hardware_name);
+    ROS_ERROR_STREAM_THROTTLE_NAMED(2, "execution_interface","No end effector controller found for " << hardware_name << ". Controllers are: " << service.response);
     return false;
   }
   if (!found_main_controller)
   {
-    ROS_ERROR_STREAM_THROTTLE_NAMED(2, "execution_interface","No main controller found for " << hardware_name);
+    ROS_ERROR_STREAM_THROTTLE_NAMED(2, "execution_interface","No main controller found for " << hardware_name << ". Controllers are: " << service.response);
     return false;
   }
 
@@ -385,7 +401,7 @@ bool ExecutionInterface::getFilePath(std::string &file_path, const std::string &
 
   // Check that the directory exists, if not, create it
   fs::path path;
-  path = fs::path(package_path_ + "/trajectories");
+  path = fs::path(package_path_ + "/trajectories/analysis/");
 
   boost::system::error_code returnedError;
   fs::create_directories( path, returnedError );
@@ -407,9 +423,23 @@ bool ExecutionInterface::getFilePath(std::string &file_path, const std::string &
 
 moveit::core::RobotStatePtr ExecutionInterface::getCurrentState()
 {
+  // Get the fake current state
+  if (unit_testing_enabled_)
+  {
+    //ROS_WARN_STREAM_NAMED("manipulation","Unit testing enabled, get current state is skipping planning scene");
+    return current_state_;
+  }
+
+  // Get the real current state
   planning_scene_monitor::LockedPlanningSceneRO scene(planning_scene_monitor_); // Lock planning scene
   (*current_state_) = scene->getCurrentState();
   return current_state_;
+}
+
+bool ExecutionInterface::enableUnitTesting()
+{
+  unit_testing_enabled_ = true;
+  return true;
 }
 
 } // end namespace
